fixed some errors

+++ src/game/server/score/logger.cpp
@@ -1,6 +1,6 @@
 #include <game/server/score/logger.h>
 
-#include <game/engine/shared/config.h>
+#include <engine/shared/config.h>
 
 #include <cstdio>
 #include <ctime>
@@ -110,8 +110,8 @@ void CScoreLogger::LogFinish(int ClientID, int Team, int StartTick)
     char Name[64], Clan[64];
     enquote(Server()->ClientName(ClientID), Name, Name + 64);
     enquote(Server()->ClientClan(ClientID), Clan, Clan + 64);
-    fprintf(gs_File, "finish %I64d \"%s\" %d %d %d %d \"%s\" \"%s\"\n",
-            UnixTime, g_Config.m_SvMap,
+    fprintf(gs_File, "finish %lld \"%s\" %d %d %d %d \"%s\" \"%s\"\n",
+            (long long)UnixTime, g_Config.m_SvMap,
             ClientID, Team,
             StartTick, CurTick,
             Name, Clan);
@@ -124,8 +124,8 @@ void CScoreLogger::LogCheckpoint(int ClientID, int Team, int StartTick, int Chec
     char Name[64], Clan[64];
     enquote(Server()->ClientName(ClientID), Name, Name + 64);
     enquote(Server()->ClientClan(ClientID), Clan, Clan + 64);
-    fprintf(gs_File, "checkpoint %I64d \"%s\" %d %d %d %d %d \"%s\" \"%s\"\n",
-            UnixTime, g_Config.m_SvMap,
+    fprintf(gs_File, "checkpoint %lld \"%s\" %d %d %d %d %d \"%s\" \"%s\"\n",
+            (long long)UnixTime, g_Config.m_SvMap,
             ClientID, Team,
             Checkpoint,
             StartTick, CurTick,
